Skill 컨트롤러 호다닥 헬퍼 제거
